# app/services/tuber-api/v1/docs/swagger.json
{
    "schemes": [
        "http",
        "https",
        "ws",
        "wss"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the TUber application.",
        "title": "TUber API",
        "contact": {},
        "version": "develop"
    },
    "host": "localhost:3000",
    "basePath": "/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Login will add a user if they do not exist or update them if they do.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Token",
                        "name": "id_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User successfully logged in",
                        "schema": {
                            "$ref": "#/definitions/authgrp.AppUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/drivers": {
            "get": {
                "description": "QueryAll will query drivers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "driver"
                ],
                "summary": "get all drivers",
                "responses": {
                    "200": {
                        "description": "Drivers successfully queried",
                        "schema": {
                            "$ref": "#/definitions/drivergrp.AppDriver"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Create will add a driver if they do not exist or update them if they do.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "driver"
                ],
                "summary": "create a new driver",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "New Driver",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/drivergrp.AppNewDriver"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Driver successfully created",
                        "schema": {
                            "$ref": "#/definitions/drivergrp.AppDriver"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/drivers/{id}": {
            "get": {
                "description": "QueryByID will query drivers by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "driver"
                ],
                "summary": "get drivers by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Drivers successfully queried",
                        "schema": {
                            "$ref": "#/definitions/drivergrp.AppDriver"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/favorite-drivers": {
            "get": {
                "description": "QueryFavorite will query favorite drivers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "driver"
                ],
                "summary": "get favorite drivers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Drivers successfully queried",
                        "schema": {
                            "$ref": "#/definitions/drivergrp.AppDriver"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/favorite-drivers/{id}": {
            "post": {
                "description": "AddFavorite will add a driver to favorite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "driver"
                ],
                "summary": "add a driver to favorite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Drivers successfully queried",
                        "schema": {
                            "$ref": "#/definitions/drivergrp.AppDriver"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "authgrp.AppUser": {
            "type": "object",
            "properties": {
                "acceptNotification": {
                    "type": "boolean"
                },
                "bio": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageURL": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "drivergrp.AppDriver": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "license": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "plate": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                }
            }
        },
        "drivergrp.AppNewDriver": {
            "type": "object",
            "required": [
                "brand",
                "color",
                "license",
                "model",
                "plate"
            ],
            "properties": {
                "brand": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "license": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "plate": {
                    "type": "string"
                }
            }
        },
...
    }
}